requirements.txt
streamlit
faiss-cpu
sentence-transformers
langchain-ollama
langchain-core
pandas
numpy
